Pipeline running without memory access errors
Pipeline runs and seems to be using correct tensors as well as
doing correct computation. However, it eventually hangs in the
lambda section and rejects all lambda requests it seems.

In addition the weight updates can show up as nan. I'm not sure if
this is a memory thing or the asynchronous computation can inherently
lead to this.