Adding bounded global bounded staleness
Major: Adding staleness bounding into NodeManager (NM). When a node has
finished all chunks in an epoch it sends a message through the NM to
all other nodes to update them on current progress. When a node gets far
enough ahed to block because of staleness, it checks the NM socket for
buffered incoming messages to see when it can move forward. Has a counter
per chunk and per node to keep track of every nodes progress.

Minor: Logging added for tracking progress of all nodes and making sure
staleness is working correctly (should remove this soon).