Implementation of global bounded staleness done
Fixed bug in which I was using the epoch % staleness to be index
into the staleness counter vectors when I needed to be using
epoch % (staleness + 1) because that is the length.

Printing the nubmer of local edges in the graph metrics. Removing
a lot of the extra logging that was for debugging.