Terminating properly by removing 'flushControl'
Removed 'flushControl()' function from destory commManager and now it terminates
properly.
However it also unfortunately needs to sleep for 5 seconds after initializintg
the data sockets to connect properly otherwise it hangs.
This was likely also the case before but we never noticed due to the
connection of the control sockets basically acting as the buffer that the
sleep is now providing.